attestation-agent: Downgrade scroll to 0.12.0
PR https://github.com/confidential-containers/guest-components/pull/1010
bumped scroll from 0.12.0 to 0.13.0, which requires rust 1.85.

This PR is a mechanical bump comming from dependabot with no code
changes. Investigation on the scroll crate repo [1] was a bit
tedious becaue of missing release tags and missing entries in the
changelog file. It doesn't seem there are any major functional
change between 0.12.0 and 0.13.0.

Based on that, downgrade scroll to 0.12.0. This is a temporary workaround
until we have rust 1.85 in our build environment.

[1] https://github.com/m4b/scroll

Signed-off-by: Greg Kurz <groug@kaod.org>