Downgrade rust edition for protos
PR #1090 introduced a new crate and tagged it with edition 2024. This
breaks the build of downstream users that cannot use rust 1.85 yet.

After double-checking with the author, it appears that the crate
doesn't strictly require the 2024 edition.

Downgrade the edition to 2021 :
- this reflects better what the crate actually requires
- this allow builds with older rust versions

Signed-off-by: Greg Kurz <groug@kaod.org>
(cherry picked from commit 52313b49730bbcbef3cc9ca1d3be4820f468aae0)
Signed-off-by: Greg Kurz <groug@kaod.org>